feature 'address_sorter': suggest and no_change RoutingMode added for UseCases, distance type changed to Float; snackbar minor fix

+++ core/domain/src/main/java/net/maxsmr/core/domain/entities/feature/address_sorter/routing/AddressRoute.kt
@@ -10,6 +10,6 @@ import net.maxsmr.core.domain.entities.feature.address_sorter.Address
 @Serializable
 class AddressRoute(
     val id: Long,
-    val distance: Long,
+    val distance: Float,
     val duration: Long?,
 )
